refactor: make __app less cluttered
