+++ project-5/components/wave/index.tsx
@@ -1,41 +1,13 @@
-import { useContext, useRef } from 'react'
+import { Canvas } from '@react-three/fiber'
 
-import { useFrame, useThree } from '@react-three/fiber'
+import Wave from './wave-mesh'
 
-import { ThemeContext } from '~contexts/theme-provider'
-
-import WaveMaterial from './wave-material'
-
-const Wave = () => {
-  const { theme } = useContext(ThemeContext)
-
-  const ref = useRef<{
-    time: number
-    noiseAmplitude: number
-    colorStart: string
-    colorEnd: string
-  }>()
-  const { width, height } = useThree((state) => state.viewport)
-
-  useFrame((state, delta) => {
-    if (ref.current) ref.current.time += delta
-  })
-
+export default function WaveWrapper() {
   return (
-    <mesh scale={[width, height, 1]}>
-      <planeGeometry />
-      {/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */}
-      {/* @ts-ignore */}
-      <waveMaterial
-        ref={ref}
-        key={WaveMaterial.key}
-        toneMapped={true}
-        colorStart={theme === 'dark' ? '#353535' : '#f7f5f2'}
-        colorEnd={theme === 'dark' ? '#000' : '#ffffff'}
-        noiseAmplitude={theme === 'dark' ? 0.5 : 2}
-      />
-    </mesh>
+    <div className="absolute top-0 left-0 w-full h-full">
+      <Canvas camera={{ position: [0, 0, 1] }}>
+        <Wave />
+      </Canvas>
+    </div>
   )
 }
-
-export default Wave
