+++ project-5/pages/_app.tsx
@@ -2,18 +2,18 @@ import { NextComponentType } from 'next'
 
 import type { AppProps } from 'next/app'
 import Head from 'next/head'
-import Image from 'next/image'
 
 import { Work_Sans } from '@next/font/google'
 import localFont from '@next/font/local'
-import { Canvas } from '@react-three/fiber'
 import { AnimatePresence } from 'framer-motion'
 
 import { Page } from 'common/types'
 import { Header, RoundedCorner } from '~components/layout'
+import Logo from '~components/logo'
 import Preloader from '~components/preloader'
-import Wave from '~components/wave'
+import WaveMesh from '~components/wave'
 import ThemeCursorProvider from '~contexts/index'
+
 import '../styles/globals.css'
 
 const latinFont = Work_Sans({ subsets: ['latin', 'latin-ext'] })
@@ -26,7 +26,8 @@ export default function App({
   pageProps,
   router: { locale, pathname, asPath },
 }: AppProps & { Component: NextComponentType & Page }) {
-  const { hasLogo = true } = Component
+  const { hasLogo = true, waveBackground } = Component
+  const { className } = locale == 'ru' ? cyrillicFont : latinFont
 
   return (
     <>
@@ -35,43 +36,21 @@ export default function App({
         <meta name="description" content="NR" />
       </Head>
 
-      <style jsx global>
-        {`
-          html {
-            font-family: ${(locale == 'ru' ? cyrillicFont : latinFont).style
-              .fontFamily};
-          }
-        `}
-      </style>
+      <main className={`${className} w-full h-full`}>
+        <ThemeCursorProvider>
+          <Preloader duration={3000} />
+          {pathname !== '/404' && <Header />}
 
-      <ThemeCursorProvider>
-        <Preloader duration={3000} />
-        {pathname !== '/404' && <Header />}
+          <RoundedCorner waveBackground={!!waveBackground}>
+            {waveBackground && <WaveMesh />}
 
-        <RoundedCorner waveBackground={!!Component.waveBackground}>
-          {Component.waveBackground && (
-            <div className="absolute top-0 left-0 w-full h-full">
-              <Canvas camera={{ position: [0, 0, 1] }}>
-                <Wave />
-              </Canvas>
-            </div>
-          )}
-
-          <AnimatePresence mode="wait">
-            {hasLogo && (
-              <Image
-                src="/nr-logo.svg"
-                width="40"
-                height="40"
-                alt="logo"
-                className="absolute mix-blend-difference top-4 left-4"
-              />
-            )}
-
-            <Component {...pageProps} key={asPath} />
-          </AnimatePresence>
-        </RoundedCorner>
-      </ThemeCursorProvider>
+            <AnimatePresence mode="wait">
+              {hasLogo && <Logo />}
+              <Component {...pageProps} key={asPath} />
+            </AnimatePresence>
+          </RoundedCorner>
+        </ThemeCursorProvider>
+      </main>
     </>
   )
 }
